Added RSM saving to bromedit. Merry Christmas Also fixed some warnings

+++ browedit/windows/ColorWindow.cpp
@@ -24,12 +24,12 @@ ColorWindow::ColorWindow(blib::ResourceManager* resourceManager, BrowEdit* browE
 	auto val = this->getComponent<blib::wm::widgets::Textbox>("value");
 	img->addDragHandler([=](int x, int y)
 	{
-		float angle = atan2(y - 128, x - 128);
+		float angle = (float)atan2(y - 128, x - 128);
 
 		glm::vec3 hsv;
 		hsv[0] = glm::mod(angle + glm::pi<float>() * 2.5f, glm::two_pi<float>()) / (glm::pi<float>() * 2) * 360;
 		hsv[1] = glm::clamp(glm::length(glm::vec2(x - 128, y - 128)) / 128.0f, 0.0f, 1.0f);
-		hsv[2] = atof(val->text.c_str());
+		hsv[2] = (float)atof(val->text.c_str());
 		this->color = blib::util::hsv2rgb(hsv);
 
 		hue->text = std::to_string(hsv[0]);
@@ -50,7 +50,7 @@ ColorWindow::ColorWindow(blib::ResourceManager* resourceManager, BrowEdit* browE
 	hue->addChangeHandler([=]()
 	{
 		glm::vec3 hsv = blib::util::rgb2hsv(this->color);
-		hsv[0] = atof(hue->text.c_str());
+		hsv[0] = (float)atof(hue->text.c_str());
 		this->color = blib::util::hsv2rgb(hsv);
 		sat->text = std::to_string(hsv[1]);
 		val->text = std::to_string(hsv[2]);
@@ -59,7 +59,7 @@ ColorWindow::ColorWindow(blib::ResourceManager* resourceManager, BrowEdit* browE
 	sat->addChangeHandler([=]()
 	{
 		glm::vec3 hsv = blib::util::rgb2hsv(this->color);
-		hsv[1] = atof(sat->text.c_str());
+		hsv[1] = (float)atof(sat->text.c_str());
 		this->color = blib::util::hsv2rgb(hsv);
 		hue->text = std::to_string(hsv[0]);
 		val->text = std::to_string(hsv[2]);
@@ -68,7 +68,7 @@ ColorWindow::ColorWindow(blib::ResourceManager* resourceManager, BrowEdit* browE
 	val->addChangeHandler([=]()
 	{
 		glm::vec3 hsv = blib::util::rgb2hsv(this->color);
-		hsv[2] = atof(val->text.c_str());
+		hsv[2] = (float)atof(val->text.c_str());
 		this->color = blib::util::hsv2rgb(hsv);
 		hue->text = std::to_string(hsv[0]);
 		sat->text = std::to_string(hsv[1]);
